finalized details screen
